# perf: reduce String allocation and use Unsafe to optimize performance