--- a/codec-http/src/main/java/io/netty/handler/codec/http/QueryStringDecoder.java
+++ b/codec-http/src/main/java/io/netty/handler/codec/http/QueryStringDecoder.java
@@ -162,7 +162,7 @@ public class QueryStringDecoder {
         }
         String rawQuery = uri.getRawQuery();
         // Also take care of cut of things like "http://localhost"
-        this.uri = rawQuery == null? rawPath : rawPath + '?' + rawQuery;
+        this.uri = rawQuery == null ? rawPath : rawPath + '?' + rawQuery;
         this.charset = checkNotNull(charset, "charset");
         this.maxParams = checkPositive(maxParams, "maxParams");
         this.semicolonIsNormalChar = semicolonIsNormalChar;
@@ -334,6 +334,14 @@ public class QueryStringDecoder {
     }
 
     private static String decodeComponent(String s, int from, int toExcluded, Charset charset, boolean isPath) {
+        if (charset.equals(CharsetUtil.UTF_8)) {
+            return decodeUtf8Component(s, from, toExcluded, isPath);
+        } else {
+            return decodeNonUtf8Component(s, from, toExcluded, charset, isPath);
+        }
+    }
+
+    private static String decodeUtf8Component(String s, int from, int toExcluded, boolean isPath) {
         int len = toExcluded - from;
         if (len <= 0) {
             return EMPTY_STRING;
@@ -361,7 +369,7 @@ public class QueryStringDecoder {
         for (int i = firstEscaped; i < toExcluded; i++) {
             char c = s.charAt(i);
             if (c != '%') {
-                strBuf.append(c != '+' || isPath? c : SPACE);
+                strBuf.append(c != '+' || isPath ? c : SPACE);
                 continue;
             }
 
@@ -370,7 +378,54 @@ public class QueryStringDecoder {
                 if (i + 3 > toExcluded) {
                     throw new IllegalArgumentException("unterminated escape sequence at index " + i + " of: " + s);
                 }
-                buf[bufIdx++] = decodeHexByte(s, i + 1);
+                PlatformDependent.putByte(buf, bufIdx++, decodeHexByte(s, i + 1));
+                i += 3;
+            } while (i < toExcluded && s.charAt(i) == '%');
+            i--;
+
+            Utf8Utils.decodeUtf8(buf, 0, bufIdx, strBuf);
+        }
+        return strBuf.toString();
+    }
+
+    private static String decodeNonUtf8Component(String s, int from, int toExcluded, Charset charset, boolean isPath) {
+        int len = toExcluded - from;
+        if (len <= 0) {
+            return EMPTY_STRING;
+        }
+        int firstEscaped = -1;
+        for (int i = from; i < toExcluded; i++) {
+            char c = s.charAt(i);
+            if (c == '%' || c == '+' && !isPath) {
+                firstEscaped = i;
+                break;
+            }
+        }
+        if (firstEscaped == -1) {
+            return s.substring(from, toExcluded);
+        }
+
+        // Each encoded byte takes 3 characters (e.g. "%20")
+        int decodedCapacity = (toExcluded - firstEscaped) / 3;
+        byte[] buf = PlatformDependent.allocateUninitializedArray(decodedCapacity);
+        int bufIdx;
+
+        StringBuilder strBuf = new StringBuilder(len);
+        strBuf.append(s, from, firstEscaped);
+
+        for (int i = firstEscaped; i < toExcluded; i++) {
+            char c = s.charAt(i);
+            if (c != '%') {
+                strBuf.append(c != '+' || isPath ? c : SPACE);
+                continue;
+            }
+
+            bufIdx = 0;
+            do {
+                if (i + 3 > toExcluded) {
+                    throw new IllegalArgumentException("unterminated escape sequence at index " + i + " of: " + s);
+                }
+                PlatformDependent.putByte(buf, bufIdx++, decodeHexByte(s, i + 1));
                 i += 3;
             } while (i < toExcluded && s.charAt(i) == '%');
             i--;