--- a/codec-http/src/main/java/io/netty/handler/codec/http/QueryStringDecoder.java
+++ b/codec-http/src/main/java/io/netty/handler/codec/http/QueryStringDecoder.java
@@ -395,7 +395,7 @@ public class QueryStringDecoder {
         for (int i = firstEscaped; i < toExcluded; i++) {
             char c = s.charAt(i);
             if (c != '%') {
-                charBuf[charBufIdx++] = c != '+' || isPath ? c : SPACE;
+                PlatformDependent.putChar(charBuf, charBufIdx++, c != '+' || isPath ? c : SPACE);
                 continue;
             }