… codec-http/src/test/java/io/netty/handler/codec/http/QueryStringDecoderTest.java
… common/src/main/java/io/netty/util/internal/PlatformDependent.java
… common/src/main/java/io/netty/util/internal/PlatformDependent0.java
… microbench/src/main/java/io/netty/handler/codec/http/QueryStringDecoderBenchmark.java