Motivation:

Reduce memory allocation and optimize performance for `QueryStringDecoder`.

Modification:

* Replace `new String()` with `Utf8Utils.decodeUtf8` when charset is UTF-8.
* Use `Unsafe` to optimize performance.
* Use `char[]` instead of `StringBuilder` to optimize performance

Result:

Before:
```
Benchmark                                         Mode  Cnt  Score   Error   Units
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  1.380 ± 0.140  ops/us
QueryStringDecoderBenchmark.noDecoding           thrpt    6  6.670 ± 0.236  ops/us
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  1.825 ± 1.031  ops/us
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  1.549 ± 0.146  ops/us
```

After:
```
Benchmark                                         Mode  Cnt  Score   Error   Units
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  1.728 ± 0.313  ops/us
QueryStringDecoderBenchmark.noDecoding           thrpt    6  6.567 ± 0.997  ops/us
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  1.892 ± 0.405  ops/us
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  2.040 ± 0.076  ops/us
```