@franz1981 2020-03-06
It looks to me that we have many changes here  and I am not sure they worth it, given that the original benchmark is already quite intensive GC-wise and it's focused on always decoding the same strings. 
In addition, beware using unsafe to save bound checks because depending on the cases you would loose any auto-vectorization chance (if there is some).
I would proceed like this:
1) cleanup/modify a bit the bench to be sure that it measures what you're expecting to stress
2) we really need that many new utf8utils methods or there is something we already have on Netty to be improved/merged/reused?

These are my 2 cents on this NS thanks for the PR and the hard work on it :)