@normanmaurer 2020-03-06
I agree with @franz1981 here... Honestly I am not sure it is worth to maintain all the code for this