@LinShunKang 2020-03-07
Add new benchmark `longMixedDecoding`, result:
Before:
```
Benchmark                                         Mode  Cnt     Score     Error   Units
QueryStringDecoderBenchmark.longMixedDecoding    thrpt    6   270.983 ±  38.867  ops/ms
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  1460.959 ± 227.085  ops/ms
QueryStringDecoderBenchmark.noDecoding           thrpt    6  7310.070 ±  62.510  ops/ms
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  1588.976 ±  42.605  ops/ms
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  1802.212 ± 508.778  ops/ms
```

After:
```
Benchmark                                         Mode  Cnt     Score     Error   Units
QueryStringDecoderBenchmark.longMixedDecoding    thrpt    6   289.256 ±  42.580  ops/ms
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  1638.161 ± 136.633  ops/ms
QueryStringDecoderBenchmark.noDecoding           thrpt    6  7159.545 ± 393.757  ops/ms
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  1644.291 ±  63.117  ops/ms
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  2161.531 ±  76.752  ops/ms
```