@LinShunKang 2020-03-08
Use char[] instead of StringBuilder to optimize performance, result from https://github.com/netty/netty/pull/10092/commits/f724ee644647f4a1d4649dbc2d454964cf02483d

Before:
```
Benchmark                                         Mode  Cnt     Score     Error   Units
QueryStringDecoderBenchmark.longMixedDecoding    thrpt    6   270.983 ±  38.867  ops/ms
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  1460.959 ± 227.085  ops/ms
QueryStringDecoderBenchmark.noDecoding           thrpt    6  7310.070 ±  62.510  ops/ms
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  1588.976 ±  42.605  ops/ms
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  1802.212 ± 508.778  ops/ms
```

After
```
Benchmark                                         Mode  Cnt     Score      Error   Units
QueryStringDecoderBenchmark.longMixedDecoding    thrpt    6   370.368 ±    3.704  ops/ms
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  2009.134 ±   23.247  ops/ms
QueryStringDecoderBenchmark.noDecoding           thrpt    6  6946.912 ± 1306.370  ops/ms
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  2204.528 ±  296.804  ops/ms
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  2690.175 ±   51.252  ops/ms
```