@LinShunKang 2020-03-08
Use `FastThreadLocal` to reduce `byte[]` and `char[]` allocation, result from https://github.com/netty/netty/pull/10092/commits/b7a374587c32aa3763dd4d4c44b204e1de931938:

Original:
```
Benchmark                                         Mode  Cnt     Score     Error   Units
QueryStringDecoderBenchmark.longMixedDecoding    thrpt    6   270.983 ±  38.867  ops/ms
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  1460.959 ± 227.085  ops/ms
QueryStringDecoderBenchmark.noDecoding           thrpt    6  7310.070 ±  62.510  ops/ms
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  1588.976 ±  42.605  ops/ms
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  1802.212 ± 508.778  ops/ms
```

After:
```
Benchmark                                         Mode  Cnt     Score     Error   Units
QueryStringDecoderBenchmark.longMixedDecoding    thrpt    6   390.395 ±  73.598  ops/ms
QueryStringDecoderBenchmark.mixedDecoding        thrpt    6  1554.062 ± 186.379  ops/ms
QueryStringDecoderBenchmark.noDecoding           thrpt    6  6183.945 ± 315.638  ops/ms
QueryStringDecoderBenchmark.nonStandardDecoding  thrpt    6  2256.158 ±  83.484  ops/ms
QueryStringDecoderBenchmark.onlyDecoding         thrpt    6  2602.271 ± 914.358  ops/ms
```