@LinShunKang 2020-03-08
@franz1981 I optimize `QueryStringDecoder` because `QueryStringDecoder` allocate too much memory:

Before optimize：
````
cn.shade.gateway.core.codec.QueryStringDecoder.decodeComponent: 7,324,670,728 (8.56%) bytes in 178,510,676 (12.38%) objects (avg size 41 bytes)
        char[]: 4,675,845,776 (5.46%) bytes in 86,376,625 (5.99%) objects (avg size 54 bytes)
        java.lang.String: 2,003,941,416 (2.34%) bytes in 83,497,559 (5.79%) objects (avg size 24 bytes)
        byte[]: 368,531,072 (0.43%) bytes in 2,879,149 (0.19%) objects (avg size 128 bytes)
        java.nio.HeapCharBuffer: 138,176,304 (0.16%) bytes in 2,878,673 (0.19%) objects (avg size 48 bytes)
        java.nio.HeapByteBuffer: 138,176,160 (0.16%) bytes in 2,878,670 (0.19%) objects (avg size 48 bytes)
````

After optimize：
````
cn.shade.gateway.core.codec.QueryStringDecoder.decodeComponent: 2,082,722,280 (2.36%) bytes in 56,660,130 (3.99%) objects (avg size 37 bytes)
        char[]: 1,402,956,216 (1.59%) bytes in 28,336,544 (1.99%) objects (avg size 50 bytes)
        java.lang.String: 679,766,064 (0.77%) bytes in 28,323,586 (1.99%) objects (avg size 24 bytes)
````

The original `QueryStringDecoder` allocated 7,324MB memory, but the optimized `QueryStringDecoder` only allocated 2,082MB memory.

PS: 
* `cn.shade.gateway.core.codec.QueryStringDecoder` is a base on Netty (before 4.1.45) `QueryStringDecoder`
* These information collected by aprof