@franz1981 2020-03-08
Thanks for the explanation.
Re
> because QueryStringDecoder allocate too much memory

Yes, but still, just once per QueryStringDecoder: the benchmark is promoting a non realistic usage of it, just for the sake of measuring the decoding cost (and I don't agree on it TBH, I would have written the benchmark differently).
If your use case shows that QueryStringDecoder::parameters is a bottleneck for your use case I suggest to focus on reduce the amount of garbage produced without the need of that many changes anyway (maybe trying to understand why many intermediate allocations are not correctly escaped?).