@LinShunKang 2020-03-08
These information collected by aprof, it was use wrk rather than JMH to benchmark, and all http requests were realistic.

And 
> maybe trying to understand why many intermediate allocations are not correctly escaped?

Sorry, I don't understand these words.